Additions for ORA2 instructor override score feature.
ORA2 has been modified in PR #7 to include functionality
allowing the instructor to override a student's ORA score.

This override score is stored in Submissions as a Score without
a related submission. API methods have been added to create and
get the override score.

The process by which these PR's should be reviewed/merged is laid
out in ORA PR Stanford-Online/edx-ora2#7